requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
audioop-lts==0.2.2
beautifulsoup4==4.13.5
blinker==1.9.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
cloudpathlib==0.21.1
colorama==0.4.6
comm==0.2.2
comtypes==1.4.12
confection==0.1.5
contourpy==1.3.1
curl_cffi==0.13.0
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
executing==2.2.0
feedparser==6.0.11
filelock==3.19.1
fonttools==4.55.0
frozendict==2.4.6
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
gnews==0.4.2
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.71.2
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
inflection==0.5.1
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.3.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
multitasking==0.0.12
murmurhash==1.0.13
narwhals==2.1.2
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==2.1.3
openai==1.107.3
packaging==24.2
pandas==2.2.3
parso==0.8.4
peewee==3.18.2
pillow==11.0.0
platformdirs==4.3.8
plotly==6.3.0
preshed==3.0.10
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.0
pypiwin32==223
python-dateutil==2.9.0.post0
pyttsx3==2.99
pytz==2024.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
Quandl==3.6.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.3
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
setuptools==80.9.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.3
srsly==2.5.1
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
streamlit==1.49.1
sympy==1.13.3
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.4
toml==0.10.2
torch==2.8.0+cpu
torchaudio==2.8.0+cpu
torchvision==0.23.0+cpu
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.4
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.2
uritemplate==4.2.0
urllib3==2.2.3
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
websockets==15.0.1
wordcloud==1.9.4
wrapt==1.17.3
yfinance==0.2.65
